// src/interfaces/ReverseGeolocationResult.ts
// Generated by https://quicktype.io

export interface ReverseGeolocationResult {
	latitude: number;
	longitude: number;
	continent: string;
	lookupSource: string;
	continentCode: string;
	localityLanguageRequested: string;
	city: string;
	countryName: string;
	countryCode: string;
	postcode: string;
	principalSubdivision: string;
	principalSubdivisionCode: string;
	plusCode: string;
	locality: string;
	localityInfo: LocalityInfo;
}

export interface LocalityInfo {
	administrative: Ative[];
	informative: Ative[];
}

export interface Ative {
	name: string;
	description: string;
	isoName?: string;
	order: number;
	adminLevel?: number;
	isoCode?: string;
	wikidataId?: string;
	geonameId?: number;
}